+++ config/application.rb
@@ -35,8 +35,6 @@ module Homs
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
 
-    config.asset_symlink = %w(hbw.css)
-
     config.allow_concurrency = true
 
     config.action_controller.permit_all_parameters = true