+++ config/webpack/loaders/babel.js
@@ -1,0 +1,12 @@
+module.exports = {
+  test:    /\.(js|jsx|ts|tsx)?$/,
+  exclude: /node_modules/,
+  use:     [{
+    loader:  'babel-loader',
+    options: {
+      cacheDirectory:   true,
+      cacheCompression: false,
+      compact:          false
+    }
+  }]
+};